+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.29 (pre-release):
 * [Enhancement] Support game's UI scale in all dialogs.
+* [Enhancement] Reduce the EVA canister dry mass to 1kg.
 * [Fix] Total mass is not ajdusted when fuel is expended from the EVA tank.
 * [Fix] Recalculate the resources mass on load to fix wrong state of the EVA tank.
 * [Fix #394] Refuel tank throws error on refuel action.