+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 1.29 (pre-release):
+# 1.29 (January 4th, 2022):
 * [Enhancement] Support game's UI scale in all dialogs.
 * [Enhancement] Reduce the EVA canister dry mass to 1kg.
 * [Fix] Total mass is not ajdusted when fuel is expended from the EVA tank.